+++ src/OpenTelemetry.Instrumentation.AspNet/Implementation/HttpInListener.cs
@@ -17,7 +17,6 @@
 using System;
 using System.Diagnostics;
 using System.Web;
-using System.Web.Routing;
 using OpenTelemetry.Context.Propagation;
 using OpenTelemetry.Internal;
 using OpenTelemetry.Trace;
@@ -26,8 +25,7 @@ namespace OpenTelemetry.Instrumentation.AspNet.Implementation;
 
 internal sealed class HttpInListener : IDisposable
 {
-    private readonly PropertyFetcher<object> routeFetcher = new("Route");
-    private readonly PropertyFetcher<string> routeTemplateFetcher = new("RouteTemplate");
+    private readonly HttpRequestRouteHelper routeHelper = new();
     private readonly AspNetInstrumentationOptions options;
 
     public HttpInListener(AspNetInstrumentationOptions options)
@@ -133,26 +131,7 @@ internal sealed class HttpInListener : IDisposable
                 activity.SetStatus(SpanHelper.ResolveActivityStatusForHttpStatusCode(activity.Kind, response.StatusCode));
             }
 
-            var routeData = context.Request.RequestContext.RouteData;
-
-            string? template = null;
-            if (routeData.Values.TryGetValue("MS_SubRoutes", out object msSubRoutes))
-            {
-                // WebAPI attribute routing flows here. Use reflection to not take a dependency on microsoft.aspnet.webapi.core\[version]\lib\[framework]\System.Web.Http.
-
-                if (msSubRoutes is Array attributeRouting && attributeRouting.Length == 1)
-                {
-                    var subRouteData = attributeRouting.GetValue(0);
-
-                    _ = this.routeFetcher.TryFetch(subRouteData, out var route);
-                    _ = this.routeTemplateFetcher.TryFetch(route, out template);
-                }
-            }
-            else if (routeData.Route is Route route)
-            {
-                // MVC + WebAPI traditional routing & MVC attribute routing flow here.
-                template = route.Url;
-            }
+            var template = this.routeHelper.GetRouteTemplate(context.Request);
 
             if (!string.IsNullOrEmpty(template))
             {